requirements.txt
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
ecdsa==0.17.0
fastapi==0.70.0
h11==0.12.0
idna==3.3
joblib==1.1.0
numpy==1.21.3
pandas==1.3.4
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
threadpoolctl==3.0.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
